Added "Keep Me Logged In"
"Keep Me Logged In" toggle in the login screen. Stores a hashed version of the users ID on the server in a database, and on the client as a cookie along with their user ID.

When the client connects, the server will check if the cookie stored on the client matches the one that's in the database. If it does, it logs the user in and regenerates a brand new hash stored in the database and on the client.