Referer Based Buttons
Fixed a bug with the Go Back button on the mod details page. Now when the framework loads it checks if the users HTTP_REFERER is from the a4a-mods.com website. If it is, the link in the Go Back button becomes the HTTP_REFERER. Otherwise the button is hidden.

On any view file, you can now use `@a4a_referer` to check if the referer is the A4A website.