Bug Fix, New Tab External Links, Linkable headers
Fixed bug where version/log text was flipped in mod details page
Links in mod descriptions to external sites now open in a new tab
It's now possible to link to specific headers in mod descriptions, by using a slugified version of the header text. E.G "This main header" would be linked by using "#this-main- header" in a link.